+++ src/main/java/com/jcloisterzone/ai/DelayedServer.java
@@ -108,10 +108,6 @@ public class DelayedServer implements RmiProxy {
         server.placeLittleBuilding(lbType);
     }
 
-    public void chatMessage(Integer player, String message) {
-        server.chatMessage(player, message);
-    }
-
     @Override
     public void undo() {
         server.undo();