+++ src/main/java/com/jcloisterzone/rmi/ClientStub.java
@@ -10,7 +10,9 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
+import javax.naming.directory.NoSuchAttributeException;
 import javax.swing.JOptionPane;
 import javax.swing.SwingUtilities;
 
@@ -21,6 +23,7 @@ import com.jcloisterzone.Expansion;
 import com.jcloisterzone.config.Config.AutostartConfig;
 import com.jcloisterzone.config.Config.DebugConfig;
 import com.jcloisterzone.config.Config.PresetConfig;
+import com.jcloisterzone.event.ChatEvent;
 import com.jcloisterzone.event.ClientListChangedEvent;
 import com.jcloisterzone.event.setup.ExpansionChangedEvent;
 import com.jcloisterzone.event.setup.PlayerSlotChangeEvent;
@@ -38,6 +41,7 @@ import com.jcloisterzone.wsio.Connection;
 import com.jcloisterzone.wsio.MessageDispatcher;
 import com.jcloisterzone.wsio.MessageListener;
 import com.jcloisterzone.wsio.WsSubscribe;
+import com.jcloisterzone.wsio.message.ChatMessage;
 import com.jcloisterzone.wsio.message.ClientListMessage;
 import com.jcloisterzone.wsio.message.ErrorMessage;
 import com.jcloisterzone.wsio.message.GameMessage;
@@ -50,6 +54,7 @@ import com.jcloisterzone.wsio.message.SlotMessage;
 import com.jcloisterzone.wsio.message.StartGameMessage;
 import com.jcloisterzone.wsio.message.TakeSlotMessage;
 import com.jcloisterzone.wsio.message.WsMessage;
+import com.jcloisterzone.wsio.server.RemoteClient;
 
 
 public class ClientStub  implements InvocationHandler, MessageListener {
@@ -57,6 +62,7 @@ public class ClientStub  implements InvocationHandler, MessageListener {
     protected final transient Logger logger = LoggerFactory.getLogger(getClass());
 
     private Connection conn;
+    private RemoteClient[] remoteClients;
     private MessageDispatcher dispatcher = new MessageDispatcher();
 
     private Game game;
@@ -117,6 +123,15 @@ public class ClientStub  implements InvocationHandler, MessageListener {
             //game.post(new PhaseEnterEvent(phase));
         }
         game.flushEventQueue();
+    }
+
+    public RemoteClient getClientById(String clientId) {;
+        for (RemoteClient remote: remoteClients) {
+            if (remote.getClientId().equals(clientId)) {
+                return remote;
+            }
+        }
+        throw new NoSuchElementException();
     }
 
 
@@ -204,7 +219,13 @@ public class ClientStub  implements InvocationHandler, MessageListener {
 
     @WsSubscribe
     public void handleClientList(ClientListMessage msg) {
+        remoteClients = msg.getClients();
         game.post(new ClientListChangedEvent(msg.getClients()));
+    }
+
+    @WsSubscribe
+    public void handleChat(ChatMessage msg) {
+        game.post(new ChatEvent(getClientById(msg.getClientId()), msg.getText()));
     }
 
     @WsSubscribe