+++ src/main/java/com/jcloisterzone/wsio/message/ClientListMessage.java
@@ -1,15 +1,15 @@
 package com.jcloisterzone.wsio.message;
 
 import com.jcloisterzone.wsio.WsMessageCommand;
-import com.jcloisterzone.wsio.server.Connection;
+import com.jcloisterzone.wsio.server.RemoteClient;
 
 @WsMessageCommand("CLIENT_LIST")
 public class ClientListMessage implements WsMessage {
 
     private String gameId;
-    private Connection[] clients;
+    private RemoteClient[] clients;
 
-    public ClientListMessage(String gameId, Connection[] clients) {
+    public ClientListMessage(String gameId, RemoteClient[] clients) {
         super();
         this.gameId = gameId;
         this.clients = clients;
@@ -23,11 +23,11 @@ public class ClientListMessage implements WsMessage {
         this.gameId = gameId;
     }
 
-    public Connection[] getClients() {
+    public RemoteClient[] getClients() {
         return clients;
     }
 
-    public void setClients(Connection[] clients) {
+    public void setClients(RemoteClient[] clients) {
         this.clients = clients;
     }